requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.2.1
certifi==2025.1.31
click==8.1.8
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.2.2
fastapi==0.103.2
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.5
MarkupSafe==2.1.5
orjson==3.9.7
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==0.21.1
python-multipart==0.0.8
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==1.4.46
starlette==0.27.0
typing_extensions==4.7.1
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.15.0
